+++ backend/Storage.cs
@@ -1,10 +1,10 @@
 public static class Storage
 {
-    public static  Dictionary<string,double> Exposicoes = new()
+    public static  Dictionary<string,decimal> Exposicoes = new()
     {
         {"PETR4", 0},
         {"VALE3", 0},
-        {"VIIA4" , 0 }
+        {"VIIA4" , 0}
 
     };
